update resource from float32 to float64

+++ core/types/receipt.go
@@ -4,8 +4,8 @@ import "github.com/ecoball/go-ecoball/common"
 
 type Receipt struct {
 	Hash   common.Hash
-	Cpu    float32
-	Net    float32
+	Cpu    float64
+	Net    float64
 	Result []byte
 }
